ooojj

+++ Proyecto_1_Grupo/Program.cs
@@ -12,8 +12,8 @@ namespace Proyecto_1_Grupo
         {
             public string modelo;
             public int anio;
-            public long kmActual;
-            public long kmService;
+            public int kmActual;
+            public int kmService;
             public string color;
             public string duenio;
         }
